Reconfigured routing to handle multiple apps
