+++ entities/admins/controller.js
@@ -3,12 +3,13 @@ const passport = require('passport');
 const jwt = require('jsonwebtoken');
 
 const setUserInfo = require('./helper').setUserInfo;
+const roleAuth = require('./helper').roleAuth;
+const roles = require('./helper').roles;
 
 module.exports = function(app, config) {
 
   // Model and Config
   var Admin = require('./model');
-  var User = require('../users/model');
 
   const adminRoutes = express.Router();
   const authRoutes = express.Router();
@@ -17,6 +18,8 @@ module.exports = function(app, config) {
   app.use('/admin', adminRoutes);
   adminRoutes.use('/auth', authRoutes);
   adminRoutes.use('/api', apiRoutes);
+
+  require('./api')(apiRoutes, config);
 
   // Middleware to require login/auth
   const requireLogin = passport.authenticate('admin', {session: false});
@@ -37,7 +40,8 @@ module.exports = function(app, config) {
     });
   });
 
-  authRoutes.post('/register', function(req, res, next) {
+  authRoutes.post('/register', requireLogin, roleAuth(roles.ROLE_ADMIN),
+  function(req, res, next) {
     if (!req.body.username || !req.body.password) {
       return next('Bad Registration: Could not find username and password');
     }
@@ -81,9 +85,6 @@ module.exports = function(app, config) {
       });
     });
   });
-
-  // Data
-  require('./api')(apiRoutes, config);
 
   // Index
   adminRoutes.get('/*', function response(req, res) {
