+++ entities/users/api.js
@@ -1,0 +1,5 @@
+const passport = require('passport');
+
+module.exports = function(routes, config) {
+
+};
