+++ static/app/components/apps/User/pages/auth/actions/types.js
@@ -1,0 +1,6 @@
+// Auth
+export const AUTH_USER = 'auth_user',
+  UNAUTH_USER = 'unauth_user',
+  AUTH_ERROR = 'auth_error',
+  FORGOT_PASSWORD_REQUEST = 'forgot_password_request',
+  RESET_PASSWORD_REQUEST = 'reset_password_request';
