+++ static/app/proptypes/index.js
@@ -38,10 +38,10 @@ export const User = {
   outcomeStmt: PropTypes.string.isRequired,
   firstHackathon: PropTypes.bool.isRequired,
   shirtSize: PropTypes.string.isRequired,
-  diet: PropTypes.string.isRequired,
+  diet: PropTypes.string,
   food: PropTypes.string,
-  website: PropTypes.string.isRequired,
-  github: PropTypes.string.isRequired,
+  website: PropTypes.string,
+  github: PropTypes.string,
   year: PropTypes.number.isRequired,
   university: PropTypes.string.isRequired,
   phone: PropTypes.number.isRequired,
