+++ static/app/static/Cookies.js
@@ -1,0 +1,10 @@
+export default {
+  admin: {
+    token: 'adminToken',
+    user: 'adminUser'
+  },
+  user: {
+    token: 'token',
+    user: 'user'
+  }
+};
