+++ views/pages/user/home.pug
@@ -1,0 +1,5 @@
+//- Application page in React
+extends ../../layouts/react
+
+block title
+    title SD Hacks 2017
